[#15] 리뷰반영 - refactor: 컴포넌트 파일 명 파스칼 케이스 적용 - feat: SearchedPlace.tsx에 onClick props 구현
