+++ src/app/signUp/page.tsx
@@ -1,4 +1,4 @@
-import SignUp from "@/app/signUp/signUp";
+import SignUp from "@/app/signUp/SignUp";
 
 export default function Page() {
   return <SignUp />;
